Удалены файлы связанные с registration, весь функционал из удаленных файлов перенесён в User файлы + возвращен конструктор в UserDataSet (Отсутствие приводило к ошибке)
