+++ .idea/sqldialects.xml
@@ -2,7 +2,6 @@
 <project version="4">
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/src/main/java/ru/mail/park/dao/impl/AuthorizationDAOImpl.java" dialect="MySQL" />
-    <file url="file://$PROJECT_DIR$/src/main/java/ru/mail/park/dao/impl/RegistrationDAOImpl.java" dialect="MySQL" />
     <file url="file://$PROJECT_DIR$/src/main/java/ru/mail/park/dao/impl/UserDAOImpl.java" dialect="MySQL" />
   </component>
 </project>
