--- src/main/java/ru/mail/park/dao/RegistrationDAO.java
@@ -1,11 +1,0 @@
-package ru.mail.park.dao;
-
-import org.springframework.http.ResponseEntity;
-
-/**
- * Created by zac on 21.10.16.
- */
-
-public interface RegistrationDAO {
-    ResponseEntity registration(String jsonString);
-}
