+++ src/main/java/ru/mail/park/dao/UserDAO.java
@@ -9,6 +9,8 @@ import javax.servlet.http.HttpSession;
  */
 
 public interface UserDAO {
+    ResponseEntity registration(String jsonString);
+
     ResponseEntity getUserInfo(long userId, HttpSession httpSession);
 
     ResponseEntity changeUserInfo(long userId, String jsonString, HttpSession httpSession);
